Game camera, animations, enemy
I have added a game camera that follows the player in a "rpg diablo-ish"
way.  I have programmed local animations they are yet to be implemented
over network. Then i have added an enemy that is triggered within a
sertain range and will run towards the player with animation and then
stop in a proper distance. I have also added basic attack animation on
the player and when it hits the enemy life will withdraw from the target
enemy.